Moving utils outside tests
