+++ client/src/test/helper.ts
@@ -49,18 +49,3 @@ export async function setTestContent(content: string): Promise<boolean> {
   )
   return editor.edit(editBuilder => editBuilder.replace(all, content))
 }
-
-export async function allWollokFiles(): Promise<Uri[]> {
-  return workspace.findFiles('**/*.{wlk,wtest,wpgm}')
-}
-
-export function replaceAll(string: string, search: string | RegExp, replace: string): string {
-  return string.toString().split(search).join(replace)
-}
-
-export function asOSString(string: string): string {
-  if (process.platform === 'win32') {
-    return replaceAll(string, '"', '\\"')
-  }
-  return string
-}
